Move DOM code to separate module
